requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
google-genai==1.7.0
numpy>=1.20.0
requests>=2.25.0 